+++ app/tests/evaluation_tests/test_config.py
@@ -2,7 +2,7 @@ import pytest
 
 from grandchallenge.components.models import InterfaceKindChoices
 from grandchallenge.evaluation.models import Evaluation
-from tests.factories import EvaluationFactory, PhaseFactory
+from tests.evaluation_tests.factories import EvaluationFactory, PhaseFactory
 from tests.utils import get_view_for_user
